+++ app/src/main/java/trickyquestion/messenger/MainScreen/MainTabsContent/ContentAdapter/Holders/FriendViewHolder.java
@@ -1,21 +1,10 @@
 package trickyquestion.messenger.MainScreen.MainTabsContent.ContentAdapter.Holders;
 
 import android.app.Activity;
-import android.content.DialogInterface;
 import android.graphics.Bitmap;
-import android.graphics.Point;
-import android.graphics.Rect;
 import android.graphics.drawable.BitmapDrawable;
-import android.graphics.drawable.Drawable;
-import android.os.AsyncTask;
-import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.RecyclerView;
-import android.util.Log;
-import android.view.Display;
-import android.view.Gravity;
 import android.view.View;
-import android.view.Window;
-import android.view.WindowManager;
 import android.widget.TextView;
 import android.widget.Toast;
 
@@ -23,9 +12,7 @@ import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.OnClick;
 import de.hdodenhof.circleimageview.CircleImageView;
-import trickyquestion.messenger.MainScreen.MainTabsContent.ContentView.PhotoDialog;
-import trickyquestion.messenger.MainScreen.MainTabsContent.PhotoDialog.PhotoDialogView;
-import trickyquestion.messenger.MainScreen.View.MainActivity;
+import trickyquestion.messenger.MainScreen.View.Dialogs.PhotoDialog;
 import trickyquestion.messenger.R;
 import trickyquestion.messenger.Util.Constants;
 
@@ -72,6 +59,6 @@ public class FriendViewHolder extends RecyclerView.ViewHolder {
         photoDialog.setBackground(screenShot);
         photoDialog.show();
 
-        // TODO: 06.09.17 Add blur with content
+        // TODO: 06.09.17 Add blur with content and cut this lines to another file
     }
 }
