+++ app/src/main/java/trickyquestion/messenger/MainScreen/View/Dialogs/IDialogView.java
@@ -1,0 +1,7 @@
+package trickyquestion.messenger.MainScreen.View.Dialogs;
+
+
+public interface IDialogView {
+    void show();
+    boolean isShow();
+}
