+++ app/src/main/java/trickyquestion/messenger/MainScreen/View/Dialogs/SettingMenuDialog.java
@@ -1,19 +1,20 @@
-package trickyquestion.messenger.MainScreen.MainTabsContent.PhotoDialog;
+package trickyquestion.messenger.MainScreen.View.Dialogs;
 
 import android.app.Dialog;
 import android.content.Context;
 import android.content.DialogInterface;
+import android.util.Log;
 import android.view.Gravity;
 import android.view.Window;
 import android.widget.RelativeLayout;
 
 import trickyquestion.messenger.R;
 
-public class PhotoDialogView {
+public class SettingMenuDialog {
     private Dialog dialog;
     private boolean isShow;
 
-    public PhotoDialogView(final Context context) {
+    public SettingMenuDialog(final Context context) {
         dialog = new Dialog(context);
         create();
     }
@@ -24,8 +25,9 @@ public class PhotoDialogView {
 
     private void create() {
         dialog.requestWindowFeature(Window.FEATURE_NO_TITLE);
-        dialog.setContentView(R.layout.dialog_photo);
-        dialog.getWindow().getAttributes().gravity = Gravity.CENTER;
+        dialog.setContentView(R.layout.dialog_menu);
+        dialog.getWindow().getAttributes().gravity = Gravity.BOTTOM;
+        dialog.getWindow().getAttributes().width = RelativeLayout.LayoutParams.MATCH_PARENT;
         dialog.setOnDismissListener(new DialogInterface.OnDismissListener() {
             @Override
             public void onDismiss(DialogInterface dialogInterface) {
