+++ app/src/main/res/layout/dialog_photo.xml
@@ -7,6 +7,7 @@
         android:layout_width="220dp"
         android:layout_height="220dp"
         android:layout_centerHorizontal="true"
+        android:layout_centerVertical="true"
         android:src="@mipmap/ic_launcher"/>
 
 </RelativeLayout>
